fix(ci): keep E2E scheduler imports package-local
